early handoff preparation in auto and on button
this will ideally be combined with early auto arm rises to save time and get our final piece in.
additionally, it is interesting but not every handoff is using the handoffForPlacement, which is clearly an issue
as the handoffForPlacement gets rid of one of the intermediate steps necesary.

additionally the changes from Monday and Wednesday seem to have not been committed, hence the separate branch. PLS commit and PUSH code so that I can be a little gremlin and work on it from home. :)